+++ dart/lib/src/http_client/failed_request_client.dart
@@ -1,4 +1,5 @@
 import 'package:http/http.dart';
+import 'sentry_http_client_error.dart';
 import '../protocol.dart';
 import '../hub.dart';
 import '../hub_adapter.dart';
@@ -126,10 +127,15 @@ class FailedRequestClient extends BaseClient {
           requestDuration: stopwatch.elapsed,
         );
       } else if (failedRequestStatusCodes.containsStatusCode(statusCode)) {
+        final message =
+            'Event was captured because the request status code was $statusCode';
+        final httpException = SentryHttpClientError(message);
+
         // Capture an exception if the status code is considered bad
         await _captureEvent(
+          exception: exception ?? httpException,
           request: request,
-          reason: failedRequestStatusCodes.toDescription(),
+          reason: message,
           requestDuration: stopwatch.elapsed,
         );
       }
@@ -144,12 +150,12 @@ class FailedRequestClient extends BaseClient {
 
   // See https://develop.sentry.dev/sdk/event-payloads/request/
   Future<void> _captureEvent({
-    Object? exception,
+    required Object? exception,
     StackTrace? stackTrace,
     String? reason,
     required Duration requestDuration,
     required BaseRequest request,
-  }) {
+  }) async {
     // As far as I can tell there's no way to get the uri without the query part
     // so we replace it with an empty string.
     final urlWithoutQuery = request.url.replace(query: '').toString();
@@ -179,7 +185,7 @@ class FailedRequestClient extends BaseClient {
       throwable: throwableMechanism,
       request: sentryRequest,
     );
-    return _hub.captureEvent(event, stackTrace: stackTrace);
+    await _hub.captureEvent(event, stackTrace: stackTrace);
   }
 
   // Types of Request can be found here:
@@ -212,12 +218,6 @@ extension _ListX on List<SentryStatusCode> {
       return false;
     }
     return any((element) => element.isInRange(statusCode));
-  }
-
-  String toDescription() {
-    final ranges = join(', ');
-    return 'This event was captured because the '
-        'request status code was in [$ranges]';
   }
 }