+++ dart/lib/src/sentry_exception_factory.dart
@@ -24,7 +24,11 @@ class SentryExceptionFactory {
 
     if (throwable is Error) {
       stackTrace ??= throwable.stackTrace;
-    } else if (_options.attachStacktrace) {
+    }
+    // throwable.stackTrace is null if its an exception that was never thrown
+    // hence we check again if stackTrace is null and if not, read the current stack trace
+    // but only if attachStacktrace is enabled
+    if (_options.attachStacktrace) {
       stackTrace ??= StackTrace.current;
     }